+++ src/clients/TimeStampReceiver.cpp
@@ -10,13 +10,13 @@ void TimeStampReceiver::SetTimeStampProvider(std::shared_ptr<TimeStampProvider> 
 	m_timeStampProvider = std::move(timeStampProvider);
 }
 
-double TimeStampReceiver::GetTimeStamp()
+uint64_t TimeStampReceiver::GetTimeStamp()
 {
 	if (m_timeStampProvider)
 	{
 		return m_timeStampProvider->getTimestamp();
 	}
-	return 0.0;
+	return 0;
 }
